Update 2lnN.cpp
In this algorithm, we dont need to search the peak first. So that we can save lnN compare. However, when the current mid is not the peak, we have to divide the part that have peak. After we find the peak, we still need to compare both sides of the peak!