Create QueueWith2Stacks.java
Q. Queue with two stacks. Implement a queue with two stacks so that each queue operations takes a constant amortized number of stack operations.

1.  Stack will examine the item most recently added, which is LIFO (last in first out)
2. Queue will examine the item least recently added, which is FIFO (first in first out)

In this question, we need to use 2 stacks to create a queue.

First stack:  _ _ _ _ _
Second stack: _ _ _ _ _
In enqueue, we just push the item into the first stack. For example, we enqueue 1,2,3,4,5
First stack:  1,2,3,4,5
Second stack:_ _ _ _ _
When dequeuing, we cannot just pop the value from the first stack( cause it would be LILO). Instead, we pop the first stack and push this value into second stock
(when second stack is empty, we just pop till first stack is empty)

First stack:_ _ _ _ _ 
Second stack: 5, 4, 3, 2, 1
Then, we pop second stack. Since 1 is the last in second stack,  it would be the first one to be pop.

Therefore, we got our queue.